Fixed broken .eslintrc, changed ecmaVersion to 2017
There was a erroneous reference to babel-eslint plugin (which was not in package.json deps, so it never worked anyways). By changing ecmaVersion to 2017 we get async/await support, if you update your globally-installed eslint to latest version.

Also removed unneeded "html" plugin (which was not in package.json deps) and "jsx" feature (we dont use that).