Bitstamp: Implementation ripple withdrawal V2
Coexistent fix for Ripple withdrawal between V1 and V2 for Bitstamp.

V1 seems to not work at all, a  parameter called currency is required and I have not been able to achieve a correct value there.

I tested with: XRP, xrp, Ripple, Ripple and any good response, just the same: 

``` Select a valid choice. X is not one of the available choices. ```