improve bit2c fetch_trades
bit2c 'noraml' get trades api method is a mess.
by default (no params) it return a huge amount of trades (100,000) and data is returned from the beginning of market.
the 'since' param is also confusing and not complying to ccxt format. 
it takes 'trade id' as the 'since' param and not timestamp. so to work with it you need to iterate multiple times until reaching current time, each time saving the last 'trade id' to move forward
https://bit2c.co.il/home/api

my suggestion is to switch completely to get lasttrades which return according to the api: the 50 last trades.

if some backward comparability is a consideration:
old trade function can be called only if 'params' is not empty (which will always be the case when retrieving current trades) 
else use 'lasttrades'

a small fix to the 'sign' function is needed as 'lasttrades' is the only public get method that doesn't have a suffix of '.json'